runelite-api: add headicon enum
