+++ runelite-api/src/main/java/net/runelite/api/NPCComposition.java
@@ -48,5 +48,5 @@ public interface NPCComposition
 
 	int getSize();
 
-	int getOverheadIcon();
+	HeadIcon getOverheadIcon();
 }
