+++ runelite-api/src/main/java/net/runelite/api/Player.java
@@ -41,5 +41,5 @@ public interface Player extends Actor
 
 	boolean isFriend();
 
-	int getOverheadIcon();
+	HeadIcon getOverheadIcon();
 }
